+++ src/components/common/Button.js
@@ -1,20 +1,20 @@
 import React from 'react'
-import styled from "./common.module.css";
+import "./Button.css";
 
 const sizes = ['long-button', 'short-button', 'search-button'];
 
 const Button = ({buttonsize, content, type, disable}) => {
   const buttonSelect = () => {
     if (buttonsize === sizes[0]) {
-      return styled.long_button;
+      return 'long_button';
     }
 
     else if (buttonsize === sizes[1]) {
-      return styled.short_button;
+      return 'short_button';
     }
 
     else if (buttonsize === sizes[2]) {
-      return styled.search_button;
+      return 'search_button';
     }
   }
