+++ src/components/main/Slide.css
