+++ src/components/train/trainResult/ShowDepartInfo.js
