+++ src/components/train/trainResult/TimeTable.js
@@ -1,7 +1,7 @@
 import { TRAIN_OPTION } from "utils/Constant";
-import { makeStandardTimes } from "pages/train/utils/MakeNoticeTimes";
-import { makeCardClassNames } from "pages/train/utils/makeCardMargin";
-import 'pages/train/style/trainResult.css';
+import { makeStandardTimes } from "feature/train/MakeNoticeTimes";
+import { makeCardClassNames } from "feature/train/makeCardMargin";
+import 'pages/train/trainResult/trainResult.css';
 
 const BUS = "bus";
