+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import 'styles/common.css';
+import 'styles/default.css'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
