+++ src/pages/main/Main.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import Slide from 'components/common/Slide'
-import FooterMain from 'components/FooterMain';
+import Slide from 'components/main/Slide'
+import Footer from 'components/common/Footer';
 import "./Main.css";
 
 const Main = () => {
@@ -10,7 +10,7 @@ const Main = () => {
       <main id='train-main'>
         <Slide></Slide>
       </main>
-      <FooterMain />
+      <Footer />
     </>
   )
 }
