+++ src/pages/roadFind/result/BusResult.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react'
 import { useLocation, useParams } from 'react-router';
 
-import KakaoMap from './component/KakaoMap';
-import "./style/Map.css";
-import "./style/BusResult.css"
+import KakaoMap from 'components/roadFind/KakaoMap';
+import "pages/roadFind/Map.css";
+import "./BusResult.css"
 
 const BusResult = () => {
   const { state } = useLocation();
