+++ src/pages/train/trainResult/trainResult.css
