+++ src/styles/common.css
@@ -55,34 +55,3 @@ nav {
   align-items: center;
   font-size: 1.5rem;
 }
-
-/* * button  */
-button {
-  margin: 10px 5px;
-  background-color: #0047FF;
-  height: 50px;
-  border: none;
-  border-radius: 5px;
-  color: white;
-  cursor: pointer;
-  font-size: 1rem;
-  margin: 0 auto;
-}
-
-.search_button {
-  /* max-width: 200px; */
-  width: 100%;
-}
-
-.long_button {
-  max-width: 250px;
-  width: 90%;
-}
-
-.short_button {
-  width: 100px;
-}
-
-button:disabled {
-  background-color: #848484;
-}
