+++ src/styles/default.css
@@ -1,0 +1,20 @@
+* {
+  margin: 0 auto;
+  padding: 0;
+  font-family: 'Inter', sans-serif;
+  font-weight: 700;
+  list-style: none;
+}
+
+body {
+  max-width: 768px;
+  min-width: 200px;
+  width: 100vw;
+  color: white;
+  background-color: #252834;
+  box-sizing: border-box;
+}
+
+.point {
+  color: #0047FF;
+}
